Split JS functions, change the JSON content type to x-www-form-urlencoded
